Move tests from signer to wallet to match function
